ocr/easyocr/pyproject.toml
[project]
name = "easyocr-liteparse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi>=0.132.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
